PT-1951 Patient measurements - UI improvements PT-2015 make patient age accessible from script service, add fuzzy value to percentile endpoint
