+++ components/patient-measurements/rest/src/main/java/org/phenotips/data/rest/internal/AbstractMeasurementRestResource.java
@@ -27,8 +27,6 @@ import javax.inject.Inject;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.joda.time.Period;
-
 import net.sf.json.JSONObject;
 
 /**
@@ -42,27 +40,6 @@ public abstract class AbstractMeasurementRestResource extends XWikiResource
     /** Injected map of measurement handlers. */
     @Inject
     protected Map<String, MeasurementHandler> handlers;
-
-    /**
-     * Get the number of months corresponding to a period string.
-     *
-     * @param age the ISO-8601 period string, without leading 'P'
-     * @throws IllegalArgumentException if the age cannot be parsed
-     * @return number of months
-     */
-    public static Double convertAgeStrToNumMonths(String age) throws IllegalArgumentException
-    {
-        Period agePeriod;
-        agePeriod = Period.parse("P" + age);
-
-        Double ageMonths = 0.0;
-        ageMonths += agePeriod.getYears() * 12;
-        ageMonths += agePeriod.getMonths();
-        ageMonths += agePeriod.getWeeks() * 7 / 30.4375;
-        ageMonths += agePeriod.getDays() / 30.4375;
-
-        return ageMonths;
-    }
 
     /**
      * Generate a server response in case of error.
