+++ components/patient-measurements/rest/src/main/java/org/phenotips/data/rest/internal/DefaultMeasurementChartResourcesResourceImpl.java
@@ -20,6 +20,7 @@ package org.phenotips.data.rest.internal;
 import org.phenotips.data.rest.MeasurementChartResourcesResource;
 import org.phenotips.measurements.MeasurementHandler;
 import org.phenotips.measurements.MeasurementsChartConfiguration;
+import org.phenotips.measurements.internal.AbstractMeasurementHandler;
 
 import org.xwiki.bridge.DocumentAccessBridge;
 import org.xwiki.component.annotation.Component;
@@ -87,7 +88,7 @@ public class DefaultMeasurementChartResourcesResourceImpl extends AbstractMeasur
                     JSONObject measurements = measurementSet.getJSONObject("measurements");
                     double ageMonths;
                     try {
-                        ageMonths = AbstractMeasurementRestResource.convertAgeStrToNumMonths(age);
+                        ageMonths = AbstractMeasurementHandler.convertAgeStrToNumMonths(age);
                     } catch (IllegalArgumentException e) {
                         throw new WebApplicationException(generateErrorResponse(Response.Status.BAD_REQUEST,
                                 "Cannot parse age."));
